--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Constants.java
@@ -1,14 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-
-public class Constants {
-
-    public static final double g = 9.8;
-
-    public static class Shooter {
-        public static final double TICKS_PER_METER = 0;
-        public static final double K_P = 0;
-        public static final double K_I = 0;
-        public static final double K_D = 0;
-    }
-
-}
